Create CI check to update nav
If a .mdx file is added or deleted (outside of snippets), we need to update the nav in docs.json. In many instances, docusaurus handled this automagically, but it's all manual in mintlify.

This CI check tests if a file was added or deleted, and fails if the nav wasn't updated.