Create CI check to update nav (#1712)
If a .mdx file is added or deleted (outside of snippets), we need to
update the nav in docs.json. In many instances, docusaurus handled this
automagically, but it's all manual in mintlify.

This CI check tests if a file was added or deleted, and fails if the nav
wasn't updated.

### Note

If you add `noSidebar: true` to a file you're creating, it won't have to
be added to the sidebar.

## How to test

**General files**
- Pull the branch
- Delete an mdx file (outside of `snippets `or `custom`), commit, and
push
- Since docs.json still lists that page, the test should fail CI
- Now remove that page from the navigation array in docs.json, commit,
and push
- It should pass CI
- Undo your changes
- Try it again with adding a .mdx file (It'll fail because you haven't
added it to docs.json)

**Index files**
- Do the same process as above for an index file, like
`obs/events/index.mdx`

## How it works

It takes the navigation array from `docs.json`, flattens it into a big
string, then searches for the deleted/added file's name (with the file
extension chopped off). For example:
- If you delete /getting-started/javascript.mdx
- It looks from getting-started/javascript in the navigation array.
- It does the equal opposite if you add a file. It checks if you added
it to the navigation array