Add Jira and Linear to Verify Webhook Providers (#1726)
Hey there 👋🏻 

Jira and Linear are recent additions to the list of webhook providers
supported by the Verify Webhook traffic-policy.

This adds guides and screenshots for both, as well as adds each to the
left nav on the integrations page.

--

It seems the CI nav check expected the image files included in the
commit to exist in `docs.json` - I've included an exclusion rule in the
CI check.

---------

Co-authored-by: Shaquil Hansford <liuqahs15@gmail.com>