Adjust redirects CI check for edge cases
MDX files that are not mentioned in the navigation of `docs.json` are not rendered in prod, and do not need a redirect.

However, our current redirect verification action will still flag these files, as it simply checks if an mdx file was deleted.

This PR makes the action account for that edge case. If you delete a .mdx file that was not in docs.json, it shouldn't fail.