Adjust redirects CI check for edge cases; fix edge cases in nav check CI (#1743)
MDX files that are not mentioned in the navigation of `docs.json` are
not rendered in prod, and do not need a redirect.

However, our current redirect verification action will still flag these
files, as it simply checks if an mdx file was deleted.

This PR makes the action account for that edge case. If you delete a
.mdx file that was not in docs.json, it shouldn't fail.

## Bonus

This PR also fixes the is-nav-updated action. It also has this issue,
where it's failing if you delete a file that was never in the nav. I
could've swore we addressed that but this fixes it for sure.