Delete unnecessary files (#1785)
The repo has a ton of files we no longer use

- Old integration overview pages, which were outdated by our decision to
drop that structure. We used to have each integration in the sidebar as
a collapsed set of folders. We no longer do that.
- Components related to the old integrations overviews
- One-time scripts created to automate the process of formatting our
docs to fit mintlify's expectations. Totally unnecessary now

## How to test

- Click around to make sure pages still work in the preview
- Particularly, make sure the integrations pages are good (should be
fine, none of these deleted files were actually being used)
- Pull the branch locally and search for the file directories these
files are in. you shouldn't see them come up in any import statements in
any files. If they do, validate that the files being imported are the
ones being deleted in this PR