Demo of blue-sky IA restructure B
This restructure differs from https://github.com/ngrok/ngrok-docs/pull/1777 in that it breaks "Endpoints" into two separate sections:
- Agent Endpoints
- Cloud Endpoints

It still retains the "What's a Universal Gateway?" section, which serves as an overview of the universal gateway, including examples

The way it handles Agent Endpoints and Cloud Endpoints is that both contain a lot of the same content. Protocols and Bindings would be different pages under each section, but with the same content imported as snippets.

This shows the same content in multiple places, but the content for each matching page would be managed in one file in the repo (i.e., Agent Endpoints > Protocols > HTTP/S and Cloud Endpoints > Protocols > HTTP/S pull content from a single file).

## Benefits of this approach

This allows us to separate Agent and Cloud Endpoints entirely in the nav experience.

With the assumption that the overview pages would have cards that link to the other docs sections, navigation between the sections would be easy enough for users.

This version maintains the "What is a Universal Gateway" section because breaking it up completely would greatly expand the "Products" dropdown in the top nav, and there'd be a lot of pages that would have maybe one or two items in their sidebar if they were broken into dedicated sections (Domains for example).