Automate changelogs (#1786)
Adds workflows that automate changelog updates for
- Helm values
- K8s changelogs

We have scripts for this, but those have to be manually run. These
workflows will automatically do what those scripts do, which is check to
see if the helm values and k8s changelog have been updated, and make a
PR if they have.

## Prior art

We have an automated workflow for creating a PR for generated docs
-
https://github.com/ngrok/ngrok-docs/blob/main/.github/workflows/auto-gen-docs-pr.yml

You can see an example PR from this workflow here
- https://github.com/ngrok/ngrok-docs/pull/1790

## How to test

I'm not sure how to test this, actually. I believe we just have to merge
it and see if it works? I'm asking for help from the eng team. Cody is
on leave, but when he created the workflow for generated docs we just
had to yolo merge it before we could test it.
- [PR that created the gen docs workflow
here](https://github.com/ngrok/ngrok-docs/pull/1755)