Enforce universal capitalization of ngrokisms (DOC-502) (#1802)
This PR introduces that a new Vale rule that enforces the capitalization
of:

- Traffic Policy / Policies
- Cloud Endpoint(s)
- Kubernetes Operator(s)
- Traffic Inspector(s)
- Agent Endpoint(s)

and applies this across the board to all (non-generated) docs.

Are there other terms we definitely want to capitalize all the time? I
also considered:

- Universal Gateway - but there are valid lowercase uses in a few docs
- Operator (without Kubernetes) - and I did capitalize a ton of these,
but there are some edge cases that make a hard and fast rule annoying
- same for Secrets and Vaults

---------

Co-authored-by: Shaquil Hansford <liuqahs15@gmail.com>